requirements.txt
requests
numpypython-dotenv
